crypto: unpack-objects: unpack encrypted packfile
The first part of an encrypted packfile is the plain header, while the
other data (excluding the tailer checksum) are encrypted.  When we
calling the function `fill()`, it will decrypt all the data and mangle
the header.

So, add a new function `fill_header()` which will read and parse the
plain header and try to decrypt remaining data in buffer, if packfile
is encrypted.

Signed-off-by: Han Xin <hanxin.hx@alibaba-inc.com>
Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>