pre-send-pack: new hook to send notifications before sending pack
In order to send notifications when a user is cloning the repository,
add a new hook "pre-send-pack" before sending pack to user.

This hook is invoked just before starting creating the packfile.

Call "send_sideband()" in "run_pre_send_pack_hook()" directly, so we
don't need to allocate large buffers as arguments for it.

All output from stdout and stderr of the hook will be directed to client
using sideband. To make it simpler, we use:

    proc.stdout_to_stderr = 1;

This hook takes several arguments, which may variant in future versions
of git. So the implementation of the hook must tolerate different
arugments or even typos in arguments. E.g.:

    pre-send-pack --have=<NUM>   \
                  --filter=<NUM> \
                  --depth=<NUM>

Reviewed-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>
Signed-off-by: Xu Ziyang <yunhuai.xzy@alibaba-inc.com>