ls-refs: support ref-suffix <suffix> in 'ls-refs' command
It's proposed to let the client to support to  send
'ref-suffix <suffix>' to the server during excuting 'ls-refs' commands
in protocol v2. Multiple instances may be given and when they are
specified, only references having a suffix matching one of the provided
suffixes are displayed.

Only protocol level extension implementations are covered here and no
related changes to the refs filter ("for_each_fullref_in_prefixes")
because this function needs to be extended or refactored, so we will put
it in a later commit to make testing easy (no functional break). Furthur,
the subcommands like 'clone' or 'fetch', etc. are not implemented here
too, they will be implemented in subsequent commits.

Signed-off-by: Teng Long <dyroneteng@gmail.com>