gc: do gc if there are to many loose refs
The native automatic gc is only triggered when there are too many loose
objects or too many packfiles, but it will not trigger when too many
loose references are created. A large number of loose references may
lead to more IO operations.

To improve this, lets add a configuration "gc.autoLooseRefsLimit" which
will also trigger automatic gc when there are more than this loose
references.

Edit:

 * Jiang Xin:
   + 2022/8/10: release buf to fix memory leak in if_too_many_loose_refs().
   + 2022/8/10: ignore "*.lock" files when checking loose refs.

Signed-off-by: Han Xin <hanxin.hx@alibaba-inc.com>
Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>