http: support read netrc file from a specific path
Support setting the GIT_CURL_NETRC_FILE environment variable to provide
the path (absolute or relative) to the netrc file that curl should use.

Although the existing http.c code already support reading of netrc file,
it only supports reading from the root path, because there is no option
to set the CURLOPT_NETRC_FILE which can provide the path to cURL.

At the same time, $HTTPD_HOST is extracted separately so that the
environment variables of httpd host can be used in other test cases.

Signed-off-by: Chen Bojun <bojun.cbj@alibaba-inc.com>