black-hole: new option for fetch-pack to discard packfile
When we do git-fetch or git-clone, a pair of programs, "fetch-pack" and
"upload-pack", will execute on both the client and server side.

We add new option "--black-hole" and "--black-hole-verify" to "fetch-pack",
in order to discard the received pack file from remote server. This
feature is used only for git test.

The difference between these two options "--black-hole" and
"--black-hole-verify" is whether verify the received pack data or not.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>