Fix the generate test script get wrong file
The test file must start with 't', four numbers, and '-', so the regex
was wrong, just update it as 't/(t[0-9]{4,}-[^/]*.sh'.

About the '{4,}', in current, git test just have four numbers, whether
it will increase to 5 numbers or more in the future is unclear, so here
is a limit of 4 numbers minimum.

Signed-off-by: Jiuyang Xie <jiuyang.xjy@alibaba-inc.com>