Add FindLastLineFromEndAndReplace for read file end
The FindLastLineFromEndAndReplace function will read the file end line, if
the end line le '\r' or '\n', it will continue to be read until the line
with correct content is read, and compare with old string, if equal,
then will append new string, and remove after new string contents.

It is good for read big file, we will read the patches file end, so
there are many of files, this method will be more efficient.

Signed-off-by: Jiuyang Xie <jiuyang.xjy@alibaba-inc.com>