Fix cannot generate patches bug
When clone a new agit repo, it just have master branch, patchwork will
report an error:"the topic 'xxx' local and remote are inconsistent,
please....". This is because the invalid conditions, the patchwork
return the error and exit.

We have fixed it, if local doesn't have branch, patchwork will try to
use remote.

Signed-off-by: Jiuyang Xie <jiuyang.xjy@alibaba-inc.com>