direct-to-packed-refs: enable by default if fetch too many refs
The fetch to packed-refs behavior will be turn on automatically if there
are lots of references (equal to or greater than 10). You can also
disable it and use the traditional fetch to loose references behavior
by calling:

    git fetch --no-write-packed-refs ...

We can change the default "10 references" threshold by providing new git
config "pack.refStoreThreshold".

The "pack.refStoreThreshold" config variable will be used in "git fetch"
command, and maybe used in "git update-ref --stdin" in future. If the
value is a boolean "true", will set a default number "10". If the value
is false or number "0", will use the traditional write to loose refs
behavior.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>