direct-to-packed-refs: FIXME: add temporary workaround for t5510-fetch
We enable fetch-to-packed-refs and atomic-fetch by default for git-fetch
and this leads to a D/F conflict in "t5510-fetch".

TODO: We should make a fix for atomic git-fetch for this case and
contribute it upstream. E.g.: fix this workaround by splict a D/F
conflict transaction and execute the conflict reference updates one by
one, especially the delete refrence operations.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>