patchwork: solve the situation that patches have the same name
The currently generated patches do not contain serial numbers by
default, so there is the possibility of patch files with the same name.
The submission is used to temporaily solve this problem. If there is a
file with same name, an underscore and a serial number will be added
after the file, for example:"file_name_1.patch, file_name_2.patch".

Signed-off-by: Jiuyang Xie <jiuyang.xjy@alibaba-inc.com>