ls-tree: improving cohension in the print code
We use 'show_tree_long()' and 'show_tree_default()' to print entries
when we want the output in 'default' or 'default long' formats.

Function 'show_tree_common_default_long()' prints the "path" field as
the last part of output, the previous part which separately implements
in 'show_tree_long()' and 'show_tree_default()'.

We can package the separate implementation together by the extension of
"size_text" in struct "show_tree_data". By improving the cohesion in
these two locations, some benefits such as uniform processing of the
output can be achieved in future.

Signed-off-by: Teng Long <dyroneteng@gmail.com>