[wip] attach: introduce `add` subcommand
The `git attach` built-in command is used to create, update, delete,
view, download attachments on a commit, or other operations that may be
supported in the future.

This commit only take the first step. We introduced a new builtin `git
attach` before, then bring the `add` subcommand to parse the given
regular files into ATTACH_BLOBs. We then construct an ATTACH_TREE which
groups these ATTACH_BLOBs together. Next step, a ATTACH_COMMIT is built
to referenece to the ATTACH_TREE. At last, we associate the
ATTACH_REFERENCE(`refs/attachments/commits` by default) to the
ATTACH_COMMIT.

An attachment refers to an regular file. if the specified pathspec is
a dir, or a multi-level dir, without any regular files in it, we will
give warnings and do nothing with ATTACH_REFERENCE.

Signed-off-by: Teng Long <dyroneteng@gmail.com>